requirements.txt
# Core dependencies
python-telegram-bot>=20.0
httpx>=0.24.0
yt-dlp>=2023.1.1
ffmpeg-python>=0.2.0
Pillow>=9.5.0
aiosqlite>=0.19.0
aiocache>=0.12.0
PyYAML>=6.0

# Utilities
python-dotenv>=1.0.0
